Move rollback functions to its own file
