+++ src/lib.rs
@@ -16,5 +16,6 @@ pub mod vm;
 pub use opcode::Opcode;
 pub use state::VMState;
 pub use vm::EraVM;
+mod rollbacks;
 pub mod world;
 use zkevm_opcode_defs::Opcode as Variant;
